Tweaking idle/max expression in plugins
While discussing idle/max_run expressions, a corner case where reqidle could
be greater than maxrunning.  I made sure this never happens.  I also added
a comment in the code to clarify the random shuffle.  We may need to
revisit this in the future, but I think it should work for now.