Packaging SL5.5's version of tarfile.py as glideinwms_tarfile.py
tarfile.py is packaged with the system python distribution.  The problem is that there is a bug in the module that prevents the creation of "in-memory" tarballs.  The version of python that comes with SL5.5 contains a fixed tarfile.py module.  This module actually works on all python versions from 2.4.3 up to and including 2.7.  At least it does with the versions I have at my disposal.

I changed the tarSupport module to import from the glideinwms_tarfile.py module rather than the system tarfile.py module.  I updated the unittests to test the creation of "in-memory" tarballs as well.