Added new environment variable specifying the location where condor should write the ssh key for ec2 submissions
This location is the root credential directory for the frontend user that is doing the submits.  I chose this location since it *is* a credential, it is considered secure enough for user credentials, and, since condor manages these files, condor will clean them up automatically.

I also deleted the extensions directory since this has nothing to do with the glideinwms codebase.  The contents of this directory have been moved to a new repo (gwms-cloud-vms) on the same server.