Bug: #1902 Change to USE_MATCH_AUTH (True) as  the default configuration.
Bug: #1902
The match_authorization option is removed from the Factory, Submit and
VOFrontend ini files.  If present, they will be ignored.
The USE_MATCH_AUTH attribute will still be created in the xml configuration
files of both the Factory and Frontend by default but will always be set to
True.  This allows a user to remove this feature if they so choose.
In v3 this may be removed as an option entirely.

Conflicts:

	doc/components/pool_install.html