Remove FactoryMatchAttrs -- it's unused.
From Igor's 7/11/11 e-mails:

"
Hi guys.

I just noticed that we never use
FactoryMatchAttrs
in the code!!!!

While we do use JobMatchAttrs in condor_q,
we never implemented the equivalent for the condor_status toward the factory.

I cannot think of a good reason why we are doing it this way; seems like a bug to me.
The net result is that we are loading much more than needed the factory collector, wasting network bandwidth and frontend resources (both memory and cpu).

Thoughts?

Igor

PS: This is definitely not a showstopper for v2.5.2, since it has been working like this for a while.
"

"
BTW: I ran a test comparing run times with and without filtering.

For CMS, the matching phase takes about 15% longer without FactoryMatchAttrs filtering.
(77s vs 66s)

Igor
"