Fixed broken invalidation of classads in the factory and frontend when the services shut down. Bug #2084.
The factory and frontend were relying on TargetType to make sure the right classad was deadvertized.  This value is not used on the classad types Master and License which we use for our classads.  Added the classad type (GlideinMyType) to the Requirements line and a method to deadvertize the frontend globals.