Changes for Bug #2265 - allowing --ssh option to have optional value.
The --ssh option can now have an optional value specifying the user to use.
An additional change was made to the condor init.d script created to correct
for a better return code when executing the 'stop'.
If you ask to stop Condor and it is not running, it will now return 0.
It was incorrectly returning 1.