Feature #1890 - Clean directories on installs based on the instance name.
Tough one to exlpain.
This change allows you to install a new instance of the factory and frontend
without requiring the old instance to be cleaned out.  Prior to this, the
ini installer required that the install_location be completely empty.  This
was required by the basic create function of both services.
.
It turned out that I had maybe taken the easy (ignorant) way out initially.
The ini installer now follows the naming convention used by the 'create'
functions of both the factory and frontend services for the working, logs and
web directories created using the following attributes in ini/cfg xml files:
For the factory,  it is at the glidein  element glidein_name  attribute level
For the frontend, it is at the frontend element frontend_name attribute level
.
Like I said... confusing.  Not sure how to describe it better.
Bottom line is, that if you desire to retain the monitoring/staging areas of
a previous installation, all you need to do is change the instance_name in the
ini file and you can do a fresh installation w/o it being wiped out.