Feature #2375 Added a --create-template option to the manage-glidein script.
This option allows you to create an ini file template for installing a single
service.  It contains all the required attributes for that service.  It should
be understood that many of the validations that would normally insure a
working installation are bypassed since those validations are normally
performed on the node for those services.
.
This will also facilitate evaluating options that can be optional in the future.