Ignore require_voms_proxy if GLIDEIN_Glexec_Use=NEVER
With GLIDEIN_GLEXEC_USE=NEVER, the frontend should ignore require_voms_proxy,
since the user job proxy will never be seen by the site.  Changed this to add
this GLIDEIN_GLEXEC_USE to the descript files.  The glideinFrontendElement
parses it and then ignores require_voms_proxy if it is NEVER.

Also removed Conflicts:
	creation/lib/cvWParamDict.py
	frontend/glideinFrontendElement.py