Feature 2612: Frontend advertises monitoring url
Monitoring url is now specified in the frontend xml
and is transfered to the frontend descript and
advertized in the classad.

UPDATED:  This value is now optional and defaults to
stage web url with stage replaced by monitor.  The
code has been changed so that the interface FrontendDescript
is unchanged with the exception of a new function to modify
the monitoring web url.

This has been tested in the RPM version and will be tested
later in the standalone tarball version [needed since RPM
does not merge xml].