Bug #2709 - problem with multiple schedds and job_queue.log . A JOB_QUEUE_LOG attribute was introduced in Condor 7.7.5 .. ticket 2598 https://condor-wiki.cs.wisc.edu/index.cgi/tktview?tn=2598     http://research.cs.wisc.edu/condor/manual/v7.7/3_3Configuration.html#16343
Prior to the introduction of this feature a job_queue.log was always
maintained in the spool directory of each schedd.  With this change, it
appears (either a bug or by desire), the job queue log of each additional
schedd must be defined explicitly.
  SCHEDD.SCHEDDJOBS2.JOB_QUEUE_LOG = $(SCHEDD.SCHEDDJOBS2.SPOOL)/job_queue.log

If not explicitely stated, only 1 job_queue.log is used.  Hence, all jobs are
assigned to all schedd queues on a restart.