Bug 2631: frontend downtime affecting other frontends
Currently, the code sets the entry into downtime
once it finds a security class / frontend into
downtime.  Then, later frontends will not be
able to send glideins (reqidle is set to zero somewhere)

This exactly matches Igor's description of the problem.

I have corrected this issue, but this should still be
investigated to make sure nothing else should be changed
as a result.