Feature 3128 - Migrate installers to use subprocess module. . Modified the ini installer to use the new subprocessSupport.py module in place of the deprecated commands, system, spawn and popen python modules. . The subprocessSupport.py module was also modfied for the following: 1. additional useShell argument in the method to allow the command to be    executed through the shell.  This was needed in the cases where pipes,    wildcards and other shell features were needed.  It previously was set    to False thus not allowing this capability.  Added to the doc for this. . 2. Fixed a minor problem with the calling of the CalledProcessError    Exception where the exitStatus was always being changed to zero. . 3. Also fixed a minor problem in the CalledProcessError class where it    never output stderr from the call.
Note: the q/a installer was not modified at this time.
      A separate ticket should be generated for this.

Conflicts:

	install/manage-glideins
	install/services/Condor.py
	install/services/VDTClient.py
	install/services/VOFrontend.py
	install/services/common.py
	lib/subprocessSupport.py

The lib/subprocessSupport.py module is being added at this time. It should
not conflict with the changes made in v2plus back in June 2012 which
created this module but has never been brought into the master branch.
I cannot find the ticket that created it back then.  The module was created to
eliminate the popen and other modules that are being deprecated.
.
Additionally changes have been made in the all modules to show the full
path within glideinswms on the imports where needed.