GlideinWMS Feature 3169: Adding Status Total to glidefactoryclient ad
Feature 3169: Factory could advertise total jobs running per entry to frontend

It seems that the glidefactory already advertizes this information in
attributes such as "GlideinMonitorTotalStatusRunning".  I have
added all these to be advertized in the glidefactoryclient ad as well.

Turns out that the glideresource classad automatically picks this up,
so this information is now in attributes such as:
"GlideFactoryMonitorStatusRunning"

Next will be putting this into the monitoring system.