Bug #3437 Ini installer broken in master
Removed need for ldapMonitor.
This was used to query bdii for entry points.  Accessing bdii  returned way
too many entries for a user to select from so it was basically commented
out back in 2010.
The import for the ldapMonitor module it used is causing failures if
the python-ldap module is not installed.