Better VOFE counting for partitionable slots - Bug #3438
The partitionable slot is always idle, even when all of its resources
have been claimed by subslots.  The VOFE should only count it towards
the idle count when all the subslots have been reclaimed.

It's tempting to report the slot as fractionally idle if some but not all
of the subslots have been claimed, or report integer values probabalistically.
I'll resist that temptation.