Join on newline character rather than an empty string
Since the move from popen2.Popen calls to subprocess calls,
we no longer use readlines.  Instead we use splitlines on the
output returned.  A readlines call preserves the newline
characters in the output.  A splitlines call with no arguments
doesn't.  I believe that this is the only place where
preserving the newline character matters.  So we will join on
newline rather than the old empty string.