Bug fixes and workarounds to problems encountered during testing
Uncovered an interesting bug in HTCondor.  If you specify both ec2_user_data and ec2_user_data_file, HTCondor is supposed to append the contents of the file specified by ec2_user_data_file to the contents of ec2_user_data.  In our use case we pass a path to a gzip'd file in ec2_user_data_file.  HTCondor chokes on the contents and puts garbage in instead.  No errors are logged anywhere, and nothing stops the submission process.  The work around is to base64 encode the gzip'd contents.