4588: Improved logging for missing FE credentials
We were just throwing exceptions into FE logs when security credentials
didn't match.  Looks like we used to do better before multi-advertising,
but we didn't catch the errors very well after that.  We now emit
a warning to the logs if the gliderequest doesn't get advertised because
of missing or mismatched credentials.