4612: Enable entry-level glidein cleaning to proceed if FE-level fails
If a Frontend security class hold limit is hit,
glideFactoryLib.keepIdleGlideins calls sanitizeGlideins for each existing
frontend -- and it tells the main loop not to call sanitizeGlideins at
the entry level. If a frontend drops out, it may leave many held jobs behind.
(The behavior is non-intuitive -- if all FEs drop out for a security class,
sanitizeGlideins doesn't get called at the FE level, but it doesn't
veto the entry-level sanitizeGlideins!)

This patch only vetos the entry-level sanitizeGlideins if the FE-level
sanitize tried to do something.