4612: Improve glidein cleaning for a frontend security class
When checking to see if we need to submit work on behalf of a frontend
credential, we check the max_held limit for its frontend security class.
If he hit the limit, we called sanitizeGlideins, but only for that
particular credential.

This change calls sanitizeGlideins for the entire frontend security class.