Improved verbosity and notifications in failures to acquire lock files during startup
Old processes may remain stuck (e.g. on IO) or the lock files may remain in incosistent.
Files may have to be cleared or processes may have to be killed by the administrator
sometime even with a reboot.
The previous message was confusing (AlreadyRunning)
The current is more explicit on the failure to acquire the lock, specifies the lock
file and logs also the exception in the log file