6626# Add support for enabling the shared_port daemon.
There is a new knob, USE_SHARE_PORT that triggers it.
It is processed by setup_firewall.sh, a new script loaded by the glideins.
I have moved USE_CCB in it as well, and out of collector_setup.sh, as it does not need to be process that early in the execution chain.