Modify condor shutdown extression to work also with dynamic slots (ticket v3/5359)
The new expression uses attributes defined in condor 8.2. If these are not available the expression
falls back to the previous one (BC82 in the expression).
With the new attributes (TotalJobs) are defined the behavior (AC82) is close to the one specified
in the documentation:
- if no job is assigned the shutdown is triggered after GLIDEIN_Max_Idle
- if jobs are assigned and the glidein goes back to idle (ExpectedMachineGracefulDrainingCompletion),
the shutdown is triggered after GLIDEIN_Max_Tail
The expression should work also with old classads but I had no condor pre 7.5 to test it