7799: prevent glideinFactoryEntryGroup writing to entry logs
The glideinFactoryEntryGroup was opening and writing to the logs
for the glidein entry points. There can be hundreds (or thousands)
of these, and there's no reason why the group needs to write to
them all. It potentially can hit open file limits and limits
on select (for when the children fork to do work).