Changed DictFile to use binary files (bytes) and convert the strings in the dictionaries
Since DictFiles values may include text but also content of binary files, they must be treated as bytes. Dictionaries and keys though are regular string, so the interface is set when writing to the file or stream (BytesIO). The current version seem not to crash with stack trace but does not complete in reasonable time. Further work is needed.