1) Modify class GSIRequest of condorSecurity.py to allow IDTOKENS authentication by default and not to throw an exception if X509_USER_PROXY has not been set in environment
2) Fix apparent bug in restore() where it attempts to delete items
from os.environ without checking if they are set first

3) Change all classes to inherit from object so they are ready
to be ported to python 3 :):):)

NB I started to write a test case for these classes but I
would prefer to delete the entire class. Only parts of it are
actually used, the rest has unclear purpose which makes it
difficult to determine if correctly implemented.  As changed
it allows XML configurations with fewer proxies and certs than
previously.