(WIP) ProxyResourceAllocation
Add a ProxyResourceAllocation frontend plugin that takes project Resource Allocation data from OSG Topology and creates credentials based on the job's ProjectName and the project's LocalAllocationIDs.

Added an `allocation_id` field to glideinFrontendInterface.Credential that holds the LocalAllocationID, if present; this will end up as the `AllocationId` attribute in the job.

TODO: we need the topology ResourceGroup of the factory entries somehow.