Added mechanism for conditional download of files
When a file basename starts with "gconditional_" (e.g. gconditional_FEATURE_myfile), then the name after that prefix is used to search for an attribute starting with the GLIDEIN_USE_ prefix (e.g. GLIDEIN_USE_FEATURE) and the file is downloaded only if the attribute is defined in glidein_config and not empty