Don't set LD_LIBRARY_PATH in condor_startup.sh for job or daemons
The HTCondor binaries set RUNPATH in the elf header to point to a path, relative to the binaries, where their shared libraries live.  As such, we don't need to set LD_LIBRARY_PATH for the daemons.  (Aside, as the condor_startup.sh is today, it sets LD_LIBRARY_PATH to a directory where the shared libraries haven't existed for many HTCondor releases). More importantly, we really don't want to set LD_LIBRARY_PATH in the STARTER_JOB_ENVIRONMENT condor config setting, which passes that environment variable along to the job.  This is breaking LIGO singularity jobs because the STARTER_JOB_ENVIRONMENT's LD_LIBRARY_PATH is overriding the container's LD_LIBRARY_PATH, and the latter is the one we want.