Added APPTAINER/SINGULARITY_CACHEDIR and APPTAINER/SINGULARITY_TMPDIR to make sure that Apptainer/Singularity files are in the Glidein tree
the default cache is ~/.apptainer/cache which was in outside the Glidein space. Temp dir uses $TMPDIR which is $GLIDEIN_LOCAL_TMP_DIR in Glideins. The neww defaults are $GLIDEIN_LOCAL_TMP_DIR/apptainer/cache and $GLIDEIN_LOCAL_TMP_DIR/apptainer/tmp

Define the variables to an empty value to use Apptainer defaults